+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2016-04-20
+
+- Default nonlinearity for `CategoricalMLPPolicy` changed to `tanh` as well, for consistency.
+- Add `flatten_n`, `unflatten_n` support for `Discrete` and `Product` spaces.
+
 # 2016-04-17
 
 - Restored the default value of the `whole_paths` parameter in `BatchPolopt` back to `True`. This is more consistent with previous configurations.